+++ sdk/src/androidTest/java/com/bugsnag/android/AppDataTest.java
@@ -45,6 +45,7 @@ public class AppDataTest {
     @After
     public void tearDown() throws Exception {
         Async.cancelTasks();
+        client.getOrientationListener().disable();
     }
 
     @Test