+++ sdk/src/androidTest/java/com/bugsnag/android/BeforeRecordBreadcrumbsTest.java
@@ -37,6 +37,7 @@ public class BeforeRecordBreadcrumbsTest {
     @After
     public void tearDown() throws Exception {
         Async.cancelTasks();
+        client.getOrientationListener().disable();
     }
 
     @Test