+++ sdk/src/androidTest/java/com/bugsnag/android/ClientConfigTest.java
@@ -36,6 +36,7 @@ public class ClientConfigTest {
     @After
     public void tearDown() throws Exception {
         Async.cancelTasks();
+        client.getOrientationListener().disable();
     }
 
     @Test