+++ sdk/src/androidTest/java/com/bugsnag/android/ConcurrentCallbackTest.java
@@ -29,6 +29,7 @@ public class ConcurrentCallbackTest {
     @After
     public void tearDown() throws Exception {
         Async.cancelTasks();
+        client.getOrientationListener().disable();
     }
 
     @Test