+++ sdk/src/androidTest/java/com/bugsnag/android/ErrorReportApiClientTest.java
@@ -29,6 +29,7 @@ public class ErrorReportApiClientTest {
     @After
     public void tearDown() throws Exception {
         Async.cancelTasks();
+        Bugsnag.getClient().getOrientationListener().disable();
     }
 
     @SuppressWarnings("deprecation")