+++ sdk/src/androidTest/java/com/bugsnag/android/MetaDataTest.java
@@ -39,6 +39,7 @@ public class MetaDataTest {
     @After
     public void tearDown() throws Exception {
         Async.cancelTasks();
+        client.getOrientationListener().disable();
     }
 
     @Test