+++ features/steps/build_steps.rb
@@ -175,3 +175,8 @@ Then("the stacktrace in request {int} contains native frame information") do |re
     assert_not_nil(frame['lineNumber'], "The lineNumber of frame #{index} is nil")
   end
 end
+
+Then(/^the payload field "(.+)" is greater than (\d+)(?: for request (\d+))?$/) do |field_path, int_value, request_index|
+  observed_value = read_key_path(find_request(request_index)[:body], field_path)
+  assert(observed_value > int_value)
+end