+++ CHANGELOG.md
@@ -13,13 +13,21 @@
 * Add binary architecture of application to payload
  [#389](https://github.com/bugsnag/bugsnag-android/pull/389)
 
+* Make config.metadata publicly accessible
+ [#406](https://github.com/bugsnag/bugsnag-android/pull/406)
+
+* Prevent errors from leaving a self-referencing breadcrumb
+ [#391](https://github.com/bugsnag/bugsnag-android/pull/391)
+
+* Prevent Bugsnag.init from instantiating more than one client
+ [#403](https://github.com/bugsnag/bugsnag-android/pull/403)
+
 ## 4.9.3 (2018-11-29)
 
 ### Bug fixes
 
 * Handle null values in MetaData.mergeMaps, preventing potential NPE
  [#386](https://github.com/bugsnag/bugsnag-android/pull/386)
-
 
 ## 4.9.2 (2018-11-07)